refactor(#17): extract toJpeg to src/utils/imageConversion.ts
- Move image-conversion logic (canvas, downscaling, JPEG encoding) out of
  the private useOcrCapture scope and into a standalone exported utility
- toJpeg(file, maxDimension?) now accepts an optional maxDimension override,
  making it independently testable and reusable
- useOcrCapture imports toJpeg from the new utility; hook return API unchanged
- Add 7 unit tests covering: JPEG output, dimension preservation, downscaling,
  canvas failure paths, custom maxDimension, and URL revocation
