+++ frontend/src/utils/imageConversion.ts
@@ -1,0 +1,42 @@
+/** Default maximum dimension (px) for the longest side — keeps uploads fast for OCR. */
+const DEFAULT_MAX_DIMENSION = 2048;
+
+/**
+ * Convert any browser-readable image file to JPEG via Canvas.
+ * Handles HEIC (iPhone default), WebP, BMP, etc.
+ * Also downscales large images so the longest side does not exceed maxDimension.
+ */
+export function toJpeg(file: File, maxDimension = DEFAULT_MAX_DIMENSION): Promise<File> {
+  return new Promise((resolve, reject) => {
+    const url = URL.createObjectURL(file);
+    const img = new Image();
+    img.onload = () => {
+      URL.revokeObjectURL(url);
+      let { width, height } = img;
+      if (width > maxDimension || height > maxDimension) {
+        const scale = maxDimension / Math.max(width, height);
+        width = Math.round(width * scale);
+        height = Math.round(height * scale);
+      }
+      const canvas = document.createElement("canvas");
+      canvas.width = width;
+      canvas.height = height;
+      const ctx = canvas.getContext("2d");
+      if (!ctx) { reject(new Error("Canvas not supported")); return; }
+      ctx.drawImage(img, 0, 0, width, height);
+      canvas.toBlob(
+        (blob) => {
+          if (!blob) { reject(new Error("JPEG conversion failed")); return; }
+          resolve(new File([blob], "photo.jpg", { type: "image/jpeg" }));
+        },
+        "image/jpeg",
+        0.85,
+      );
+    };
+    img.onerror = () => {
+      URL.revokeObjectURL(url);
+      reject(new Error("Could not load image"));
+    };
+    img.src = url;
+  });
+}
